rename: entities 파일 API 관련 파일 명시적으로 변경
