+++ src/entities/event/model/eventDetail.ts
